refactor: move descriptions search vector to package data model
Once derivations are clustered into packages, the description lives on the package.

Migration 0027 is updated to use the historical model registry.
Otherwise removing a field from the current model breaks a fresh migration run.
